+++ src/view/main/main.tsx
@@ -8,8 +8,10 @@ import ItemRow from '../../lib/list/item_row';
 
 // Main Creators
 import { SelectCreator, selectOptions } from '../form/select';
+import { PeriodSelectorCreator, periodSelectorOptions } from '../date_time/period_selector';
 import { IconButtonCreator, iconButtonOptions } from '../button/icon_button';
 import { InputSetCreator, inputSetOptions } from '../form/input_set';
+import { SwitchSetCreator, switchSetOptions } from '../form/switch_set';
 import { DropDownMenuCreator, dropDownMenuOptions } from '../menu/dropdown_menu';
 import { PDFViewerEmbeddedCreator, pdfViewerEmbeddedOptions } from '../pdf_viewer/pdf_viewer_embedded';
 import { PDFViewerFullScreenCreator, pdfViewerFullScreenOptions } from '../pdf_viewer/pdf_viewer_full_screen';
@@ -28,7 +30,9 @@ export default function MainPage(props: {}): JSX.Element {
         "Input Set": { title: "Input Set", data: inputSetOptions(), creator: InputSetCreator },
         "PDF Viewer Embedded": { title: "PDF Viewer Embedded", data: pdfViewerEmbeddedOptions(), creator: PDFViewerEmbeddedCreator },
         "PDF Viewer Fullscreen": { title: "PDF Viewer Fullscreen", data: pdfViewerFullScreenOptions(), creator: PDFViewerFullScreenCreator },
-        "Select": { title: "Select", data: selectOptions(), creator: SelectCreator },                
+        "Select": { title: "Select", data: selectOptions(), creator: SelectCreator },
+        "Period Selector": { title: "Period Selector", data: periodSelectorOptions(), creator: PeriodSelectorCreator },
+        "Switch Set": { title: "Switch Set", data: switchSetOptions(), creator: SwitchSetCreator },
     }
 
     let content: React.ReactNode = null;